[#4335] target new style + inline editing
